requirement.txt
alembic==1.4.2
bcrypt==3.1.7
cffi==1.14.0
click==7.1.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
